fixed some small texture bleed

+++ src/main/java/useless/dragonfly/model/block/data/FaceData.java
@@ -6,7 +6,7 @@ import java.util.Arrays;
 
 public class FaceData {
 	@SerializedName("uv")
-	public float[] uv = null;
+	public double[] uv = null;
 	@SerializedName("texture")
 	public String texture = null;
 	@SerializedName("cullface")
